python-project-83/pyproject.toml
[tool.poetry]
name = "hexlet-code"
version = "0.1.0"
description = "page alalyzer project"
authors = ["Anastasia <anastasiia1803@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10,<4.0"
gunicorn = "^21.2.0"
flask = "^3.0.2"
python-dotenv = "^1.0.1"
psycopg2-binary = "^2.9.9"
validators = "^0.23.2"
beautifulsoup4 = "^4.12.3"
requests = "^2.31.0"

[tool.poetry.group.dev.dependencies]
flake8 = "^7.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
